Restructured project to separate audio and server modules. Renamed desktop application window form name.
