+++ AudioManager/Program.cs
@@ -1,4 +1,4 @@
-namespace DAIRemote
+namespace AudioManager
 {
     internal static class Program
     {
@@ -10,14 +10,8 @@ namespace DAIRemote
         {
             // To customize application configuration such as set high DPI settings or default font,
             // see https://aka.ms/applicationconfiguration.
-            Application.EnableVisualStyles();
-            Application.SetCompatibleTextRenderingDefault(false);
             ApplicationConfiguration.Initialize();
-            Application.Run(new DAIRemoteApplicationUI());
-
-
-
+            Application.Run(new AudioOutputForm());
         }
-
     }
 }
