+++ DAIRemote/DAIRemote.csproj
@@ -11,7 +11,9 @@
   
 
   <ItemGroup>
+    <ProjectReference Include="..\..\..\..\..\source\repos\AudioManager\AudioManager.csproj" />
     <ProjectReference Include="..\DisplayProfileManager\DisplayProfileManager.csproj" />
+    <ProjectReference Include="..\UDPServer\UDPServerManager.csproj" />
   </ItemGroup>
 
   <ItemGroup>
