+++ DAIRemote/DAIRemoteApplicationUI.resx
